More wpull refactoring
